requirements.txt
pandas~=2.2.3
icecream~=2.1.4
requests~=2.32.3
tqdm~=4.67.1
beautifulsoup4~=4.13.3
numpy~=2.2.3
chromadb~=0.6.3
pillow~=11.1.0
streamlit~=1.43.0
python-dotenv~=1.0.1
langchain-core~=0.3.41
uvicorn~=0.34.0
fastapi~=0.115.11
pydantic~=2.10.6
langchain-openai~=0.3.8